Implement build mode in the Makefile
Similarly to the gmp vs. nogmp build, the maintainers used two almost
identical files with differing suffixes -- which had to be copied to
the target file that was used by the rest of the system -- to implement
two build modes "ddnnf" and "bdg". Since the difference for the source
code was just the presence of the DDNNF_FULL macro, this is now done
in the Makefile with the BUILD_MODE variable. As before, this defaults
to ddnnf mode.