Avoid scientific notation in #SAT output
The number of satisfying assignments is maintained as a probability
internally which leads to floating point numbers when reporting the
answer to #SAT in the end. These in turn are printed in scientific
notation by cout when they are large enough, which makes the result
useless for people interested in exact counts.

This patch converts the floating point value to mpz_t if GMP is
enabled or keeps it a floating point number otherwise (to avoid
truncation of very large numbers from double to unsigned long).
In any case, the std::fixed I/O manipulator is applied when printing
the value, so that even doubles avoid scientific notation, at the
cost of ugly fixed-width decimals for something that is actually
an integer.